requirements.txt
python-dotenv

openai

cryptography